+++ RandomSpawnPointBruh/RandomSpawnPointBruh.cs
@@ -29,7 +29,7 @@ namespace RandomSpawnPointBruh
         //Module Constants
         private const string _pluginId = "vapok.mods.RandomSpawnPointBruh";
         private const string _displayName = "RandomSpawnPointBruh";
-        private const string _version = "2.0.8";
+        private const string _version = "2.1.0";
         
         //Interface Properties
         public string PluginId => _pluginId;